# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для базы знаний с вопросами и ответами",
        "title": "Knowledge Base API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:2709",
    "basePath": "/",
    "paths": {
...
        "/tutors": {
            "get": {
                "description": "Returns list of all tutors",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tutors"
                ],
                "summary": "Get all tutors",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Tutor"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new tutor with full name and email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tutors"
                ],
                "summary": "Create new tutor",
                "parameters": [
                    {
                        "description": "Tutor data",
                        "name": "tutor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TutorSwaggerRequestBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tutor created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tutors/{id}": {
            "get": {
                "description": "Returns tutor by specified ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tutors"
                ],
                "summary": "Get tutor by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tutor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Tutor"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Tutor not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update tutor with full name and email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tutors"
                ],
                "summary": "Update tutor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tutor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tutor data",
                        "name": "tutor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TutorSwaggerRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tutor updated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete tutor by ID",
                "tags": [
                    "tutors"
                ],
                "summary": "Delete tutor by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tutor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Tutor not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Tutor": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "models.TutorSwaggerRequestBody": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                }
            }
        }
    }
}